capstone-project/requirements.txt
djangorestframework== 3.14..0
django==5.1
djangorestframework-simplejwt==5.3.1
